Crew-AI/email_ai/pyproject.toml
[project]
name = "email_ai"
version = "0.1.0"
description = "email-ai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "langchain-community>=0.0.20",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "pandas>=2.0.0",
    "langchain-google-genai>=1.0.0"
]

[project.scripts]
email_ai = "email_ai.main:run"
run_crew = "email_ai.main:run"
train = "email_ai.main:train"
replay = "email_ai.main:replay"
test = "email_ai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
